+++ examples/ImageXpressPlateExample.py
@@ -1,10 +1,9 @@
 import shutil
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions
+from faim_hcs.hcs.acquisition import PlateLayout, TileAlignmentOptions
 from faim_hcs.hcs.converter import ConvertToNGFFPlate, NGFFPlate
 from faim_hcs.hcs.imagexpress import StackAcquisition
 from faim_hcs.stitching import stitching_utils
-from faim_hcs.Zarr import PlateLayout
 
 
 def main():
