+++ src/faim_hcs/Zarr.py
@@ -1,5 +1,4 @@
 import os
-from enum import IntEnum
 from os.path import join
 from pathlib import Path
 from typing import Union
@@ -19,47 +18,8 @@ from ome_zarr.writer import (
 )
 from zarr import Group
 
+from faim_hcs.hcs.acquisition import PlateLayout, _get_row_cols
 from faim_hcs.UIntHistogram import UIntHistogram
-
-
-class PlateLayout(IntEnum):
-    """Plate layout, 96-well or 384-well."""
-
-    I96 = 96
-    I384 = 384
-
-
-def _get_row_cols(layout: Union[PlateLayout, int]) -> tuple[list[str], list[str]]:
-    """Return rows and columns for requested layout."""
-    if layout == PlateLayout.I96:
-        rows = ["A", "B", "C", "D", "E", "F", "G", "H"]
-        cols = [str(i).zfill(2) for i in range(1, 13)]
-        assert len(rows) * len(cols) == 96
-    elif layout == PlateLayout.I384:
-        rows = [
-            "A",
-            "B",
-            "C",
-            "D",
-            "E",
-            "F",
-            "G",
-            "H",
-            "I",
-            "J",
-            "K",
-            "L",
-            "M",
-            "N",
-            "O",
-            "P",
-        ]
-        cols = [str(i).zfill(2) for i in range(1, 25)]
-        assert len(rows) * len(cols) == 384
-    else:
-        raise NotImplementedError(f"{layout} layout not supported.")
-
-    return rows, cols
 
 
 def _create_zarr_plate(
